Add route check for authenticated users
