+++ src/components/LandingPage/Banner/banner.scss
@@ -26,7 +26,7 @@ $black: #000008;
     height: 100%;
     overflow: hidden;
     background: rgba(0,0,0,0.3);
-    box-shadow: 0px 0px 40px 40px rgba(0,0,0,0.3);
+    // box-shadow: 0px 0px 40px 40px rgba(0,0,0,0.3);
     align-items: center;
     padding: 4em;
     padding-top: 16em;
