+++ src/components/LandingPage/NovelOfTheWeek/novelOfTheWeek.scss
@@ -67,3 +67,14 @@ $black: #000008;
     padding: 0;
   }
 }
+
+@media(max-width: 400px){
+  .weekNovelBox {
+    .weekNovelImage {
+      img {
+        width: 170px;
+        height: 170px;
+      }
+    }
+  }
+}
