+++ src/components/helpers/AuthModal/AuthModal.scss
@@ -4,7 +4,7 @@
     align-items: center;
     z-index: 1000;
     width: 100%;
-    height: 100vh;
+    height: 100%;
     background: rgba(0, 0, 0, 0.6);
 
   
@@ -50,14 +50,15 @@
   }
   
   @media screen and (max-width: 500px) and (min-width: 320px) {
-      h4 {
-          font-size: 18px;
-      }
-      .description {
-          font-size: 14px;
-      }
+    h4 {
+      font-size: 18px;
+    }
+    .description {
+      font-size: 14px;
+    }
     .modal {
-        min-height: 280px;
+      min-height: 280px;
+      max-height: 560px;
     }
   }
