+++ src/components/helpers/Separator/Separator.scss
